requirements.txt
dj-database-url==0.5.0
Django==3.1.3
django-allauth==0.44.0
docutils==0.16
filelock==3.0.12
gunicorn==20.0.4
mccabe==0.6.1
oauthlib==3.1.0
packaging==20.4
Pillow==8.0.1
pkginfo==1.6.1
psycopg2==2.8.6
pycparser==2.20
PyJWT==2.0.0
pylint==2.6.0
pyparsing==2.4.7
pystache==0.5.4
python3-openid==3.2.0
pywin32-ctypes==0.2.0
readme-renderer==28.0
requests==2.24.0
requests-oauthlib==1.3.0
requests-toolbelt==0.9.1
rfc3986==1.4.0
sentry-sdk==0.16.2
six==1.15.0
sqlparse==0.4.1
toml==0.10.2
virtualenv==20.1.0
webencodings==0.5.1
wrapt==1.12.1
